backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0
python-multipart==0.0.9
polar-sdk==0.9.0
